:bug: fix: Fix blocked error
